Ajustei, testei e refatorei o código das funções relacionadas ao usuário. Estão todas ok.
