+++ src/routes/userRouter.ts
@@ -1,16 +1,14 @@
 import express from 'express';
-import { befriendUser } from '../endpoints/befriendUser';
-import { getUser } from '../endpoints/getUser';
-import { getUserProfile } from '../endpoints/getUserProfile';
-import { login } from '../endpoints/login';
-import { signUp } from '../endpoints/signUp';
-import { unfriendUser } from '../endpoints/unfriendUser';
+import { UserController } from '../controller/UserController';
 
 export const userRouter = express.Router();
 
-userRouter.put("/signup", signUp);
-userRouter.post("/login", login);
-userRouter.get("/profile", getUserProfile);
-userRouter.get("/:id", getUser);
-userRouter.get("/befriend", befriendUser);
-userRouter.delete("/unfriend", unfriendUser);
+const userController = new UserController();
+
+userRouter.put("/signup", userController.signUp);
+userRouter.post("/login", userController.login);
+userRouter.get("/profile", userController.getUserProfile);
+userRouter.get("/:id", userController.getUser);
+userRouter.delete("/:id/delete", userController.deleteUser);
+userRouter.post("/befriend", userController.befriendUser);
+userRouter.delete("/unfriend", userController.unfriendUser);
